pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nano-pdf"
version = "0.2.1"
description = "A CLI tool to edit PDF slides using natural language prompts, powered by Gemini 3 Pro Image"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  {name = "Gavriel Cohen", email = "gavrielcohen89@gmail.com"}
]
keywords = [
    "pdf",
    "pdf-editor",
    "gemini",
    "ai",
    "computer-vision",
    "natural-language",
    "cli",
    "slides",
    "presentation",
    "ocr"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Office/Business",
    "Topic :: Utilities",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Environment :: Console",
]
dependencies = [
    "typer[all]",
    "pdf2image",
    "pypdf",
    "pytesseract",
    "httpx>=0.27.0",
    "python-dotenv",
    "Pillow"
]

[project.urls]
Homepage = "https://github.com/fychmen/nano-pdf"
Repository = "https://github.com/fychmen/nano-pdf"
Issues = "https://github.com/fychmen/nano-pdf/issues"

[project.scripts]
nano-pdf = "nano_pdf.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["nano_pdf*"]
